+++ tests/testthat/test-pivotExpressionSet.R
@@ -8,6 +8,6 @@ test_that("multiplication works", {
   expect_s3_class(long, "tbl_df")
   expect_equal(dim(long), c(30, 29))
   expect_setequal(rownames(sub_adat), long$array_id)
-  expect_setequal(getFeatures(sub_adat), long$feature)
+  expect_setequal(getAnalytes(sub_adat), long$feature)
   skip("Add increased coverage unit tests for `pivotExpressionSet()`")
 })