+++ NEWS.md
@@ -1,3 +1,5 @@
+# SomaDataIO (development version)
+
 # SomaDataIO 5.0.0
 
 * Initial public release to GitHub!