+++ NEWS.md
@@ -1,5 +1,8 @@
 # SomaDataIO (development version)
 
+* `getFeatures()` was renamed to be less ambiguous and better align with
+  internal SomaLogic code usage. Now use `getAnalytes()`
+
 # SomaDataIO 5.0.0
 
 * Initial public release to GitHub!