+++ NEWS.md
@@ -1,5 +1,6 @@
 # SomaDataIO (development version)
 
+* New `set_rn()` drop-in replacement for `magrittr::set_rownames()`
 * `getFeatures()` was renamed to be less ambiguous and better align with
   internal SomaLogic code usage. Now use `getAnalytes()`
 * `getFeatureData()` was also renamed to `getAnalyteInfo()`